+++ grid2op/Chronics/multifolderWithCache.py
@@ -159,13 +159,24 @@ class MultifolderWithCache(Multifolder):
 
     def reset(self):
         """
-        Rebuilt the cache as if it were built from scratch. This call might take a while to process.
+        Rebuilt the cache as if it were built from scratch. 
+        This call might take a while to process.
+        
+        .. danger::
+            You NEED to call this function (with `env.chronics_handler.reset()`)
+            if you use the `MultiFolderWithCache` class in your experiments.
+            
+        .. warning::
+            If a seed is set (see :func:`MultiFolderWithCache.seed`) then
+            all the data in the cache are also seeded when this 
+            method is called.
         """
         self._cached_data = [None for _ in self.subpaths]
         self.__i = 0
         # select the right paths, and store their id in "_order"
         super().reset()
         self.cache_size = 0
+        max_int = np.iinfo(dt_int).max
         for i in self._order:
             # everything in "_order" need to be put in cache
             path = self.subpaths[i]
@@ -176,8 +187,7 @@ class MultifolderWithCache(Multifolder):
                 max_iter=self.max_iter,
                 chunk_size=None,
             )
-            if self.seed is not None:
-                max_int = np.iinfo(dt_int).max
+            if self.seed_used is not None:
                 seed_chronics = self.space_prng.randint(max_int)
                 data.seed(seed_chronics)
 
@@ -193,10 +203,10 @@ class MultifolderWithCache(Multifolder):
 
         if self.cache_size == 0:
             raise RuntimeError("Impossible to initialize the new cache.")
-
+        
         self.__nb_reset_called += 1
         return self.subpaths[self._order]
-
+    
     def initialize(
         self,
         order_backend_loads,
@@ -227,6 +237,41 @@ class MultifolderWithCache(Multifolder):
         id_scenario = self._order[self._prev_cache_id]
         self.data = self._cached_data[id_scenario]
         self.data.next_chronics()
+    
+    @property
+    def max_iter(self):
+        return self._max_iter
+    
+    @max_iter.setter
+    def max_iter(self, value : int):
+        self._max_iter = int(value)
+        for el in self._cached_data:
+            if el is None:
+                continue
+            el.max_iter = value
+    
+    def max_timestep(self):
+        return self.data.max_timestep()
+    
+    def seed(self, seed : int):
+        """This seeds both the MultiFolderWithCache
+        (which has an impact for example on :func:`MultiFolder.sample_next_chronics`)
+        and each data present in the cache.
+
+        Parameters
+        ----------
+        seed : int
+            The seed to use
+        """
+        res = super().seed(seed)
+        max_int = np.iinfo(dt_int).max
+        for i in self._order:
+            data = self._cached_data[i]
+            if data is None:
+                continue
+            seed_ts = self.space_prng.randint(max_int)
+            data.seed(seed_ts)
+        return res
         
     def load_next(self):
         self.__nb_step_called += 1