+++ grid2op/tests/test_issue_164.py
@@ -34,7 +34,7 @@ class Issue164Tester(unittest.TestCase):
     def test_issue_164(self):
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            env = grid2op.make(reward_class=Test164_Reward, test=True)
+            env = grid2op.make('rte_case14_realistic', reward_class=Test164_Reward, test=True)
         max_timestep = env.chronics_handler.max_timestep()
 
         obs = env.reset()